Update Credential Pages to use PageRoutedTabs. Add Back to Credentials tab
